+++ tests/testthat/test-get-star.R
@@ -21,27 +21,3 @@ test_that("error when send list of designs to draw_data", {
   expect_error(draw_data(list(design, design)), "Please send a single design object to the design argument, typically created using the \\+ operator.")
   
 })
-
-test_that("get_ works", {
-  dat_with_Z <- draw_assignment(design, dat)
-  
-  expect_true(all(c("Z") %in% colnames(dat_with_Z)))
-  
-  dat_sampled <- draw_sample(design, dat[c(1:75, 1:75), ])
-  expect_equal(nrow(dat_sampled), 75)
-  
-})
-
-
-test_that("no sampling does not produce error from draw_sample", {
-  
-  design <- population + potential_outcomes + inquiry
-  
-  dat <- draw_data(design)
-  
-  expect_error(draw_sample(design, data = dat), "does not include any sampling steps")
-  
-  expect_error(draw_assignment(design, data = dat), "does not include any assignment steps")
-  
-})
-